@@ -18,14 +18,13 @@ from rich.console import Console
 from rich.panel import Panel
 from rich.table import Table
 
-from biothings import config as btconfig
 from biothings.utils import es
 from biothings.utils.common import timesofar
 from biothings.utils.manager import CLIJobManager
 from biothings.utils.dataload import dict_traverse
 from biothings.utils.serializer import load_json, to_json
-from biothings.utils.sqlite3 import get_src_db
 from biothings.utils.workers import upload_worker
+import biothings.utils.inspect as btinspect
 
 
 def get_logger(name=None):
@@ -69,7 +68,7 @@ def load_plugin_managers(
     """
     Load a data plugin from <plugin_path>, and return a tuple of (dumper_manager, upload_manager)
     """
-    from biothings import config as btconfig
+    from biothings import config
     from biothings.hub.dataload.dumper import DumperManager
     from biothings.hub.dataload.uploader import UploaderManager
     from biothings.hub.dataplugin.assistant import LocalAssistant
@@ -77,7 +76,7 @@ def load_plugin_managers(
     from biothings.utils.hub_db import get_data_plugin
 
     _plugin_path = pathlib.Path(plugin_path).resolve()
-    btconfig.DATA_PLUGIN_FOLDER = _plugin_path.parent.as_posix()
+    config.DATA_PLUGIN_FOLDER = _plugin_path.parent.as_posix()
     sys.path.append(str(_plugin_path.parent))
 
     if plugin_name is None:
@@ -96,7 +95,7 @@ def load_plugin_managers(
     LocalAssistant.uploader_manager = upload_manager
 
     assistant = LocalAssistant(f"local://{plugin_name}")
-    logger.debug(assistant.plugin_name, plugin_name, _plugin_path.as_posix(), btconfig.DATA_PLUGIN_FOLDER)
+    logger.debug(assistant.plugin_name, plugin_name, _plugin_path.as_posix(), config.DATA_PLUGIN_FOLDER)
 
     dp = get_data_plugin()
     dp.remove({"_id": assistant.plugin_name})
@@ -113,7 +112,6 @@ def load_plugin_managers(
             },
         }
     )
-    breakpoint()
     p_loader = assistant.loader
     p_loader.load_plugin()
 
@@ -310,7 +308,12 @@ def do_dump_and_upload(plugin_name, logger=None):
 
 
 def process_inspect(source_name, mode, limit, merge, logger, do_validate, output=None):
-    """Perform inspect for the given source. It's used in do_inspect function below"""
+    """
+    Perform inspect for the given source. It's used in do_inspect function below
+    """
+    from biothings import config
+    from biothings.utils import hub_db
+
     VALID_INSPECT_MODES = ["jsonschema", "type", "mapping", "stats"]
     mode = mode.split(",")
     if "jsonschema" in mode:
@@ -327,14 +330,14 @@ def process_inspect(source_name, mode, limit, merge, logger, do_validate, output
     t0 = time.time()
     data_provider = ("src", source_name)
 
-    src_db = get_src_db()
+    src_db = hub_db.get_src_db()
     pre_mapping = "mapping" in mode
     src_cols = src_db[source_name]
     inspected = {}
     converters, mode = btinspect.get_converters(mode)
     for m in mode:
         inspected.setdefault(m, {})
-    cur = src_cols.findv2()
+    cur = src_cols.find()
     res = btinspect.inspect_docs(
         cur,
         mode=mode,
@@ -536,7 +539,7 @@ def get_manifest_content(working_dir: Union[str, pathlib.Path]) -> dict:
             manifest = yaml.safe_load(yaml_handle)
     elif manifest_file_json.exists():
         with open(manifest_file_json, "r", encoding="utf-8") as json_handle:
-            manifest = load_json(json_handle)
+            manifest = load_json(json_handle.read())
     else:
         logger.info("No manifest file discovered")
     return manifest
@@ -547,7 +550,7 @@ def get_manifest_content(working_dir: Union[str, pathlib.Path]) -> dict:
 ########################
 
 
-def get_uploaders(working_dir: pathlib.Path):
+def get_uploaders(working_dir: pathlib.Path) -> list[str]:
     """
     A helper function to get the uploaders from the manifest file in the working directory
     used in show_uploaded_sources function below
@@ -555,14 +558,11 @@ def get_uploaders(working_dir: pathlib.Path):
     data_plugin_name = working_dir.name
 
     manifest = get_manifest_content(working_dir)
-    upload_section = manifest.get("uploader", None)
-    upload_sections = manifest.get("uploaders", None)
-
-    if upload_section is None and upload_sections is None:
-        table_space = [data_plugin_name]
-    elif upload_section is None and upload_sections is not None:
-        table_space = [item["name"] for item in upload_sections]
-
+    uploader_section = manifest.get("uploader", None)
+    uploaders_section = manifest.get("uploaders", None)
+    table_space = [data_plugin_name]
+    if uploader_section is None and uploaders_section is not None:
+        table_space = [item["name"] for item in uploaders_section]
     return table_space
 
 
@@ -608,10 +608,15 @@ def get_uploaded_collections(src_db, uploaders):
 
 
 def show_uploaded_sources(working_dir, plugin_name):
-    """A helper function to show the uploaded sources from given plugin."""
+    """
+    A helper function to show the uploaded sources from given plugin.
+    """
+    from biothings import config
+    from biothings.utils import hub_db
+
     console = Console()
     uploaders = get_uploaders(working_dir)
-    src_db = get_src_db()
+    src_db = hub_db.get_src_db()
     uploaded_sources, archived_sources, temp_sources = get_uploaded_collections(src_db, uploaders)
     if not uploaded_sources:
         console.print(
@@ -705,10 +710,14 @@ def do_list(plugin_name=None, dump=False, upload=False, hubdb=False, logger=None
 
 
 def serve(host, port, plugin_name, table_space):
-    """Serve a simple API server to query the data plugin source."""
+    """
+    Serve a simple API server to query the data plugin source.
+    """
     from .web_app import main
+    from biothings import config
+    from biothings.utils import hub_db
 
-    src_db = get_src_db()
+    src_db = hub_db.get_src_db()
     rprint(f"[green]Serving data plugin source: {plugin_name}[/green]")
     asyncio.run(main(host=host, port=port, db=src_db, table_space=table_space))
 
@@ -759,8 +768,11 @@ def do_clean_dumped_files(data_folder, plugin_name):
 
 def do_clean_uploaded_sources(working_dir, plugin_name):
     """Remove all uploaded sources by a data plugin in the working directory."""
+    from biothings import config
+    from biothings.utils import hub_db
+
     uploaders = get_uploaders(working_dir)
-    src_db = get_src_db()
+    src_db = hub_db.get_src_db()
     uploaded_sources = []
     for item in src_db.collection_names():
         if item in uploaders:
