@@ -5,6 +5,8 @@ from xml.etree import ElementTree
 
 from elasticsearch import Elasticsearch
 from pymongo.collection import Collection
+from config import logger
+from elasticsearch.exceptions import NotFoundError
 
 
 class _Ele(NamedTuple):  # Cleanup Element
@@ -41,7 +43,31 @@ class _Ele(NamedTuple):  # Cleanup Element
         return dom.toprettyxml(indent=" " * 2)
 
 
-def find(collection, env=None, keep=3, group_by=None, return_db_cols=False, **filters):
+def find(collection, *, env=None, keep=3, group_by=None, return_db_cols=False, **filters):
+    """
+    Identify snapshots to remove or keep based on specified criteria.
+
+    This function queries a MongoDB collection to find snapshots matching the given filters,
+    groups them according to the specified grouping key(s), and determines which snapshots
+    to keep or remove based on the 'keep' parameter.
+
+    Parameters:
+    - collection (Collection): The MongoDB collection to query. Must be an instance of `pymongo.collection.Collection`.
+    - env (str, optional): The environment name to filter snapshots. Defaults to None.
+    - keep (int, optional): The number of most recent snapshots to keep in each group. Defaults to 3.
+    - group_by (str or list, optional): The key or list of keys to group snapshots by. If None, defaults to 'build_config'.
+    - return_db_cols (bool, optional): If True, returns the raw database query results instead of the structured `_Ele` element. Defaults to False.
+    - **filters: Additional keyword arguments to filter snapshots.
+
+    Returns:
+    - _Ele or list: An `_Ele` element representing the snapshots to be removed and kept, organized by groups,
+      or a list of raw database query results if `return_db_cols` is True.
+
+    Raises:
+    - NotImplementedError: If 'collection' is not an instance of `pymongo.collection.Collection`.
+    - TypeError: If 'group_by' is neither a string, list, tuple, nor None.
+    """
+
     if not isinstance(collection, Collection):
         raise NotImplementedError("Require MongoDB Hubdb.")
 
@@ -53,7 +79,12 @@ def find(collection, env=None, keep=3, group_by=None, return_db_cols=False, **fi
     groups = list(
         collection.aggregate(
             [
-                {"$project": {"build_config": "$build_config._id", "snapshot": {"$objectToArray": "$snapshot"}}},
+                {
+                    "$project": {
+                        "build_config": "$build_config._id",
+                        "snapshot": {"$objectToArray": "$snapshot"},
+                    }
+                },
                 {"$unwind": {"path": "$snapshot"}},
                 {
                     "$addFields": {
@@ -64,6 +95,8 @@ def find(collection, env=None, keep=3, group_by=None, return_db_cols=False, **fi
                 },
                 {"$replaceRoot": {"newRoot": "$snapshot.v"}},
                 {"$match": {"environment": env, **filters} if env else filters},
+                # Exclude cloud credentials
+                {"$unset": ["conf.cloud.access_key", "conf.cloud.secret_key"]},
                 {"$sort": {"created_at": 1}},
                 {"$group": {"_id": group_by, "items": {"$push": "$$ROOT"}}},
             ]
@@ -110,33 +143,75 @@ def _remove(doc, keep):
 # because SnapshotEnv.client is not async
 
 
-def delete(collection, element, envs):
+def delete(collection, element, envs, ignoreErrors=False):
     cnt = 0
     assert element.tag == "CleanUps"
     for group in element.elems:
-        for catagory in group.elems:
-            if catagory.tag == "Remove":
-                for snapshot in catagory.elems:
-                    _delete(collection, snapshot, envs)
+        for category in group.elems:
+            if category.tag == "Remove":
+                for snapshot in category.elems:
+                    _delete(collection, snapshot, envs, ignoreErrors)
                     cnt += 1
     return cnt
 
 
-def _delete(collection, snapshot, envs):
+def _delete(collection, snapshot, envs, ignoreErrors=False):
+    """
+    Delete a single snapshot from the Elasticsearch repository and update the MongoDB collection.
+
+    This helper function deletes the specified snapshot from the Elasticsearch repository and removes
+    its reference from the MongoDB 'collection'.
+
+    Parameters:
+    - collection (Collection): The MongoDB collection where snapshot metadata is stored.
+    - snapshot (_Ele): An `_Ele` element representing the snapshot to be deleted.
+    - envs (dict): A mapping of environment names to their respective clients or configurations.
+    - ignoreErrors (bool, optional): If True, ignores errors during deletion and continues processing. Defaults to False.
+
+    Raises:
+    - AssertionError: If the tag of 'snapshot' is not 'Snapshot'.
+    - ValueError: If the environment is not registered in 'envs' and 'ignoreErrors' is False, or if the snapshot does not exist in the repository.
+    - KeyError: If required keys are missing in 'snapshot.attrs'.
+    """
+
     assert snapshot.tag == "Snapshot"
 
-    if "environment" in snapshot.attrs:
-        env = snapshot.attrs["environment"]
-        client = envs[env].client
-    else:  # legacy format
-        env = snapshot.attrs["conf"]["indexer"]["env"]
-        env = envs.index_manager[env]
-        client = Elasticsearch(**env["args"])
+    try:
+        if "environment" in snapshot.attrs:
+            env = snapshot.attrs["environment"]
+            client = envs[env].client
+        else:  # legacy format
+            env = snapshot.attrs["conf"]["indexer"]["env"]
+            env = envs.index_manager[env]
+            client = Elasticsearch(**env["args"])
+    except KeyError as exc:
+        message = (
+            f"Environment '{env}' is not registered and connection details are unavailable. "
+            "Consider adding it to the hub configuration otherwise manual deletion is required."
+        )
+        if ignoreErrors:
+            logger.error(message)
+            logger.info(
+                "Ignoring error and continuing to delete snapshot '%s'", snapshot.attrs["_id"]
+            )
+            collection.update_one(
+                {"_id": snapshot.attrs["build_name"]},
+                {"$unset": {f"snapshot.{snapshot.attrs['_id']}": 1}},
+            )
+            return
+        raise ValueError(message) from exc
 
-    client.snapshot.delete(
-        repository=snapshot.attrs["conf"]["repository"]["name"],
-        snapshot=snapshot.attrs["_id"],
-    )
+    try:
+        client.snapshot.delete(
+            repository=snapshot.attrs["conf"]["repository"]["name"],
+            snapshot=snapshot.attrs["_id"],
+        )
+    except NotFoundError as exc:
+        raise ValueError(
+            f"Snapshot '{snapshot.attrs['_id']}' does not exist in the repository "
+            f"'{snapshot.attrs['conf']['repository']['name']}'. "
+            "Validate the snapshots to remove this snapshot from the database."
+        ) from exc
 
     collection.update_one(
         {"_id": snapshot.attrs["build_name"]},
@@ -176,8 +251,7 @@ def _plain_text(snapshot):
             snapshot.attrs["_id"],
             " (",
             f'env={snapshot.attrs.get("environment") or "N/A"}',
-            ", "
-            #
+            ", ",
             # "build_name" generally agrees with the snapshot _id,
             # although technically snapshots can be named anything.
             # since in most use cases, the snapshot name at least
